543 beo 8790 security enhanced tvpapi token revocation (#216) (#255)
* BEO-8790 - Added revocation of tokens for all users + udids in domain on RemoveDomain, revocation of tokens for all users and removed udid for RemoveDeviceFromDomain.

ONLY if APIs called by admin user and avoid_revocation = false and session_revocation_enabled = true

* Removed logs

* BEO-8954 fix - Response.Clear not removing the response body

* BEO-8955

* BEO-8955

* BEO-8955

* fix

* Revert "fix"

This reverts commit 002876ad41f8880105af7123048ff01e0833a202.

* fix

* Improvements

(cherry picked from commit 0554e92ae9111c5b6586c6c610ce9328b6f83794)

+++ tvpapi/TVPApi.Common/ws/Domains/DomainService.cs
@@ -205,7 +205,14 @@ namespace TVPApiServices
 
                     if (resDomain.m_oDomainResponseStatus == DomainResponseStatus.OK)
                     {
-                        AuthorizationManager.Instance.DeleteAccessToken(groupID, initObj.Token, true);
+                        if (AuthorizationManager.IsAdmin())
+                        {
+                            AuthorizationManager.Instance.RevokeSessions(groupID, initObj.Platform, resDomain.m_oDomain, domainID, sUdid);
+                        }
+                        else
+                        {
+                            AuthorizationManager.Instance.DeleteAccessToken(groupID, initObj.Token, true);
+                        }
                     }
                 }
                 catch (Exception ex)
@@ -620,7 +627,18 @@ namespace TVPApiServices
                 }
                 try
                 {
+                    Domain domain = null;
+                    if (AuthorizationManager.IsAdmin())
+                    {
+                        domain = new TVPApiModule.Services.ApiDomainsService(groupID, initObj.Platform).GetDomainInfo(initObj.DomainID);
+                    }
+
                     resDomain = new TVPApiModule.Services.ApiDomainsService(groupID, initObj.Platform).RemoveDomain(initObj.DomainID);
+
+                    if (resDomain == DomainResponseStatus.OK && AuthorizationManager.IsAdmin())
+                    {
+                        AuthorizationManager.Instance.RevokeSessions(groupID, initObj.Platform, domain);
+                    }      
                 }
                 catch (Exception ex)
                 {
